Add main.js: Try to set ids on person headers
Get all person header elements and set the name of the person they
correspond to as their id. Doesn't currently work because it seems that
the headers get loaded via AJAX or something after the page has loaded,
so we'll need to wait until they show up on the page before manipulating
them.